Add Crossplane Configuration Package definition
Crossplane Configuration Packages offer a "native" solution to package XRD definitions and compositions that can be installed in any cluster that provides the defined dependencies.

The accompanying Github Workflow builds the package and publishes it to the GHCR when a tag is pushed.

This presents a way to version crossplane resources and packages.

Closes #1.