Add initial XRD design for XStorage/Storage (#9)
This XRD design defines a global `owner` of `buckets` to be created. In
addition, those `buckets` can be `discoverable` which can be used to
create a catalog of buckets that other owners can request access to.

Furthermore, the `bucketAccessRequests` lets the `owner` define
`buckets` which they want to request either `ReadWrite` or `ReadOnly`
access to.

Lastly, the `bucketAccessGrants` lets the `owner` grant `ReadWrite` or
`ReadOnly` access to `buckets` to other owners.

Closes #7.