Composition implementation of initial XRD design (#11)
The implementation follows the initial XRD design described in #7 and
implemented in #9.

Furthermore, it implements a simple workflow to check if a
bucketAccessRequest has been granted. This uses a Kubernetes observe
object from the provider-kubernetes Provider to check if the policy
exists and attaches it to the MinIO User object if it does.

Closes #10.