Add tutorials and how-to guides for storage-minio (#15)
We follow Divio's documentation system
(https://docs.divio.com/documentation-system/) for this documentation.

The tutorial includes a step-by-step guide for new users to get
storage-minio up and running with kind from start to finish. This
includes steps to install and configure crossplane and MinIO as well as
the configuration for provider-minio and provider-kubernetes.

The how-to guides focus more on specific questions/topics such as how to
install the configuration package, how to create buckets or how to
request/grant access.

Closes #14.