Add unit tests for storage-minio (#20)
We use crossplane render to render the manifests of tests defined in the
tests/ directory which are then compared to the expected results in the
tests/expected/ directory.

We use one file name 00x-buckets.yaml for every test in order to
separate them and test specific cases.

A GitHub Action automatically runs the unit tests on every pull request
against the main branch.

Closes #12.