Add storage-aws configuration package (#24)
This configuration package includes the same functionality as the
storage-minio configuration package but for AWS S3 bucket storage and
users. It includes the same permissions workflow and can be installed as
a standalone configuration package.

The documentation has also been updated with new tutorials and how-tos.
Furthermore, unit tests are run on every PR.

Note: E2E tests are disabled for now as the GitHub Runner seems to struggle
with the installation of provider-aws-* providers in the kind cluster.

Closes #23.