Add second user for MinIO and cred rotation CronJob
The composition creates two User objects now in order to have multiple
"keys" with the same permissions. They can be rotated (deleted) with a
CronJob. An example is given in in
examples/manifests/cred-rotation.minio.yaml. Under "interval" the
validity period of the keys can be set in seconds, e.g. 86400 for one
day.