Add second credentials to MinIO and AWS and CronJob for cred rotation (#51)
The compositions now create two User/AccessKey objects in order to have
multiple "keys" with the same permissions. They can be rotated (deleted) with 
the example CronJob given in in
examples/manifests/cred-rotation.{minio,aws}.yaml. Under "interval" the
validity period of the keys can be set in seconds, e.g. 86400 for one
day and the schedule can also be adapted.